% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAlphat.r
\name{findAlphat}
\alias{findAlphat}
\title{Alphat calculator
Plots a histogram of FACs input and allows the user to select the Alphat threshold.}
\usage{
findAlphat(x, channels)
}
\arguments{
\item{x}{data frame from scsaDat}

\item{channels}{number of channels used in FACs acquisition}
}
\value{
Returns Alphat threshold value selected by the user.
}
\description{
Alphat calculator
Plots a histogram of FACs input and allows the user to select the Alphat threshold.
}
\examples{
\dontrun{
data(zf_facs)
x <- scsaDat(dat)
findAlphat(x)
}
}
\references{
Evenson DP, Larson KL, Jost LK (2002) Sperm Chromatin Structure Assay: Its Clinical Use for Detecting Sperm DNA Fragmentation in Male Infertility and Comparisons With Other Techniques. Journal of Andrology, 23, 25–43.
}
\seealso{
scsa, scsaDat, findHDS
}
