% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHDS.r
\name{findHDS}
\alias{findHDS}
\title{High DNA stainability (HDS) calculator}
\usage{
findHDS(x, channels, ...)
}
\arguments{
\item{x}{data frame from scsaDat}

\item{channels}{number of channels used in FACs acquisition}

\item{...}{additional arguments to be passed}
}
\value{
Returns the HDS threshold selected by the user.
}
\description{
Plots a histogram of FACs input and allows the user to select the HDS threshold.
}
\examples{
\dontrun{
data(zf_facs)
x <- scsaDat(zf_facs)
findHDS(x)
}
}
\references{
Evenson DP, Larson KL, Jost LK (2002) Sperm Chromatin Structure Assay: Its Clinical Use for Detecting Sperm DNA Fragmentation in Male Infertility and Comparisons With Other Techniques. Journal of Andrology, 23, 25–43.
}
\seealso{
scsa, scsaDat, findAlphat
}
