% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsa.r
\name{scsa}
\alias{scsa}
\title{Sperm Chromatin Structure Assay (SCSA) FACs Analyzer}
\usage{
scsa(x, ...)
}
\arguments{
\item{x}{square matrix of raw FACs data from a sample}

\item{...}{additional arguments to be passed.}
}
\value{
Returns a scsa object
}
\description{
Sperm Chromatin Structure Assay (SCSA) FACs Analyzer
}
\examples{
data(zf_facs)

\dontrun{
scsa(zf_facs)
# Runs user through the SCSA pipeline.
}
dat <- scsa(zf_facs,id="sample1", thresh.alphat=0.25,thresh.HDS=500, max.green=1023)
# Returns SCSA object with manual configuration

print(dat)
# Returns sample ID, DFI and HDS

stats.scsa(dat)
# Returns SCSA statistics 

plot(dat)
# Returns a plot of FACs data for the sample.
}
\references{
Evenson DP, Larson KL, Jost LK (2002) Sperm Chromatin Structure Assay: Its Clinical Use for Detecting Sperm DNA Fragmentation in Male Infertility and Comparisons With Other Techniques. Journal of Andrology, 23, 25–43.
}
\seealso{
scsaDat, scsaStats, findHDS, findAlphat
}
