% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsaDat.r
\name{scsaDat}
\alias{scsaDat}
\title{FACs data importer}
\usage{
scsaDat(x, id, max.green = dim(x)[1], channels = dim(x)[2], ...)
}
\arguments{
\item{x}{matrix of cell counts with the "green" values in rows and "red" values in columns}

\item{id}{sample ID name}

\item{max.green}{maximum value of green to include in the analysis}

\item{channels}{number of channels used in FACs acquisition}

\item{...}{additional arguments to be passed}
}
\value{
Returns a list of the frequency of cells in each green/red channel and calculates alphat
}
\description{
Imports raw data from a FACs sample.
}
\examples{
data(zf_facs)
scsaDat <- scsaDat(x=zf_facs,id="Sample 1",max.green=1023,channels=1024)
}
\references{
Evenson DP, Larson KL, Jost LK (2002) Sperm Chromatin Structure Assay: Its Clinical Use for Detecting Sperm DNA Fragmentation in Male Infertility and Comparisons With Other Techniques. Journal of Andrology, 23, 25–43.
}
\seealso{
scsa, scsaDat, findHDS, findAlphat
}
