% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsaPlot.r
\name{scsaPlot}
\alias{scsaPlot}
\title{SCSA plot}
\usage{
scsaPlot(x, title = x$id, ...)
}
\arguments{
\item{x}{list from scsaDat function}

\item{title}{title of the plot}

\item{...}{additional arguments to be passed}
}
\value{
Returns summary plots SCSA data
}
\description{
SCSA plot
}
\examples{
\dontrun{
scsaPlot(x)
}
}
\references{
Evenson DP, Larson KL, Jost LK (2002) Sperm Chromatin Structure Assay: Its Clinical Use for Detecting Sperm DNA Fragmentation in Male Infertility and Comparisons With Other Techniques. Journal of Andrology, 23, 25–43.
}
\seealso{
scsa, scsaDat, findHDS, findAlphat
}
