% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsaStats.r
\name{scsaStats}
\alias{scsaStats}
\title{SCSA statistics calculator}
\usage{
scsaStats(x, id = x$id, thresh.alphat, thresh.HDS, ...)
}
\arguments{
\item{x}{list from scsaDat output}

\item{id}{sample id}

\item{thresh.alphat}{Alphat threshold value. If not specified, a plot will be generated for the user to select the threshold.}

\item{thresh.HDS}{High DNA Staining threshold value. If not specified, a plot will be generated for the user to select the threshold.}

\item{...}{additional arguments to be passed}
}
\value{
Returns a list of SCSA statistics
}
\description{
Calcuates SCSA statistics for a sample.
}
\details{
DNA Fragmentation index (DFI) = proportion of cells above the Alphat threshold

High DNA Stainability (HDS) = proportion of cells above the HDS threshold

N = total number of cells analyzed

X.alphat = Weighted mean DFI of all cells analyzed

sd.alphat = Standard deviation of the DFI of all cells analyzed

n.comp = number of cells above the Alphat threshold

X.comp = weighted mean DFI of cells above the Alphat threshold

sd.comp = standard deviation of DFI above the Alphat threshold

thresh.alphat = user-selected Alphat threshold

thresh.HDS = user-selected HDS threshold

X.red = mean red fluorescence of all cells

X.green = mean green fluorescence of all cells
}
\examples{
data(zf_facs)
x <- scsaDat(zf_facs)
scsaStats(x,id="Sample 1",thresh.alphat=0.27,thresh.HDS=527)
# Values for SCSA thresholds can be manually entered to generate the statistics.
}
\references{
Evenson DP, Larson KL, Jost LK (2002) Sperm Chromatin Structure Assay: Its Clinical Use for Detecting Sperm DNA Fragmentation in Male Infertility and Comparisons With Other Techniques. Journal of Andrology, 23, 25–43.
}
\seealso{
scsa, scsaDat, findHDS, findAlphat
}
