% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsa.r
\name{stats.scsa}
\alias{stats.scsa}
\title{Statistics for SCSA objects}
\usage{
stats.scsa(x, ...)
}
\arguments{
\item{x}{an object of class \code{"scsa"}}

\item{...}{additional arguments to be passed}
}
\value{
Returns a data frame of summary statistics. See scsaStats for variable descriptions.
}
\description{
Statistics for SCSA objects
}
\examples{
data(zf_facs)
dat <- scsa(zf_facs,thresh.alphat=0.25,thresh.HDS=500)
stats.scsa(dat)
}
\seealso{
scsaDat, scsaStats, findHDS, findAlphat
}
