% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zf_facs.R
\docType{data}
\name{zf_facs}
\alias{zf_facs}
\title{FACs output from a Sperm Chromatin Structure assay of a zebra finch}
\format{1024 x 1024 matrix of "green" channels in rows and "red" channels in columns. The cells contain the counts of sperm cells that emitted in that "green" and "red" channel.}
\usage{
data(zf_facs)
}
\description{
FACs output from a Sperm Chromatin Structure assay of a zebra finch
}
\keyword{datasets}
